requirements.txt
albumentations==2.0.0
matplotlib==3.10.0
numpy==2.2.2
pandas==2.2.3
Pillow==11.1.0
scikit_learn==1.6.1
torch==2.5.1
torchvision==0.20.1
torchviz==0.0.3
tqdm==4.66.5
transformers==4.47.1
